Split pipeline tests into per-handler files; drop _pipeline suffix
- Break test_handler_pipelines.py into test_ec.py, test_rhea.py,
  test_chembl.py, test_clo.py, and test_efo.py
- Rename test_mesh_pipeline.py → test_mesh.py
- Rename test_umls_pipeline.py → test_umls.py
- Update all internal references and README entries
