+++ tests/pipeline/README.md
@@ -48,17 +48,21 @@ exists it is reused — `write_umls_ids()` is not called again. This means:
   GO (2 compendia via UberGraph). Adding a new vocabulary requires only adding its fixtures
   to `conftest.py` and one entry in `VOCABULARY_REGISTRY` — this file never changes.
 
-- **`test_mesh_pipeline.py`** (`pipeline`) — MeSH-specific targeted test
+- **`test_mesh.py`** (`pipeline`) — MeSH-specific targeted test
   ([issue #675](https://github.com/NCATSTranslator/Babel/issues/675)). Downloads
   `babel_downloads/MESH/mesh.nt` automatically if absent. One test: chemicals must exclude
   D05 protein subtrees (D05.500/D05.875), D08 protein subtrees (D08.811/D08.622/D08.244),
   and D12.776 — but must include D08.211 Coenzymes (NAD, Coenzyme A) and D05.374/D05.750/D05.937
   (Micelles, Polymers, Smart Materials), which are all classified as CHEMICAL_ENTITY.
 
-- **`test_umls_pipeline.py`** (`pipeline`) — UMLS-specific targeted test. Requires
+- **`test_umls.py`** (`pipeline`) — UMLS-specific targeted test. Requires
   `UMLS_API_KEY` for the initial download (or cached files). One test: chemicals must not
   contain any UMLS IDs that the protein compendium claimed (semantic type tree
   A1.4.1.2.1.7, Amino Acid/Peptide/Protein).
+
+- **`test_ec.py`**, **`test_rhea.py`**, **`test_chembl.py`**, **`test_clo.py`**,
+  **`test_efo.py`** (`pipeline`) — Output format and content checks for the EC, Rhea,
+  ChEMBL, CLO, and EFO data handlers.
 
 - **`checks/`** (`pipeline`) — Per-compendium regression assertions tied to specific GitHub
   issues, designed for test-driven development. See [Pipeline Checks](#pipeline-checks) below.
